Move action-cable-consumer to utils
